requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
requests==2.26.0
tldextract==3.1.2
twilio==7.4.0
Twisted==21.7.0
IP2Location==8.6.4
django==3.2.13
django-filter==21.1
